+++ docs/source/conf.py
@@ -146,11 +146,12 @@ myst_enable_extensions = ['amsmath', 'dollarmath']
 # to keep in the ampworks documentation. The rest is excluded.
 
 dataset_keep = [
-    'zero_time',
-    'zero_below',
     'downsample',
     'enforce_monotonic',
-    'interactive_xy_plot',
+    'interactive_plotly',
+    'interactive_bokeh',
+    'zero_below',
+    'zero_time',
 ]
 
 richres_keep = ['copy']